rename encryption_key_manager package
